backend_locadora/requirements.txt
Flask==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
